fix: components moved into folders
