+++ src/components/CardsContainer/CardsContainer.tsx
@@ -2,12 +2,12 @@
 // import { Switch, Route } from "react-router-dom"
 
 // Custom Hooks
-import { useLoader } from '../hooks/useLoader'
+import { useLoader } from '../../hooks/useLoader'
 
 // Components
-import Skills from './Skills'
-import Projects from './Projects'
-import Services from './Services'
+import Skills from '../Skills/Skills'
+import Projects from '../Projects/Projects'
+import Services from '../Services/Services'
 
 export default function CardsContainer(): JSX.Element {
   const loader = useLoader()
