+++ src/components/NavBar/NavBar.tsx
@@ -2,11 +2,11 @@ import { useState } from 'react'
 import { Link } from 'react-router-dom'
 
 // Assets
-import doubleArrowRight from '../assets/double-arrow-right.svg'
-import navBarLogo from '../assets/navbar-logo/navbar-logo.svg'
+import doubleArrowRight from '../../assets/double-arrow-right.svg'
+import navBarLogo from '../../assets/navbar-logo/navbar-logo.svg'
 // Utils
-import { bounceIn, zoomIn, fadeIn } from '../utils/animateCss/animateCss'
-import { facebook, linkedin, mailTo, twitter } from '../utils/social-links/socialLinks'
+import { bounceIn, zoomIn, fadeIn } from '../../utils/animateCss/animateCss'
+import { facebook, linkedin, mailTo, twitter } from '../../utils/social-links/socialLinks'
 
 export default function NavBar(): JSX.Element {
   // State that handles the icons to be seen
