+++ src/components/_index.scss
@@ -1,1 +1,2 @@
-@import "./BtnToTop/BtnTotop";
+@import './BtnToTop/BtnTotop';
+@import './AboutResume/AboutResume';
