+++ src/utils/animateCss/animateCss.ts
@@ -1,11 +1,11 @@
-// esto debe ser agregado siempre para tener animado el elemento
+// this must always be added to have the element animated
 const animated: string = 'animate__animated'
 
-// velocidades de la animacion
+// animation speeds
 const fast: string = 'animate__fast'
 const faster: string = 'animate__faster'
 
-// estas son las animaciones que debes usar importandolas y usando el nombre de la variable
+// these are the animations that you should use importing them and using the variable name
 const fadeIn: string = `${animated} animate__fadeIn`
 const bounceIn: string = `${animated} animate__bounceIn ${fast}`
 const zoomIn: string = `${animated} animate__zoomIn ${faster}`
@@ -13,6 +13,5 @@ const fadeInRightBig: string = `${animated} animate__fadeInRightBig ${faster}`
 const slideInLeft: string = `${animated} animate__slideInLeft ${faster}`
 const slideUp: string = `${animated} animate__slideInUp ${faster}`
 
-/* Aqui solo añade el nombre de la animacion para exportarla 
-y poder usarla en diferentes archivos */
+// Here just add the name of the animation to export and be able to use it in different files
 export { fadeIn, bounceIn, zoomIn, fadeInRightBig, slideInLeft, slideUp }
