+++ src/views/HomePage.tsx
@@ -1,8 +1,8 @@
 // components
-import Hero from '../components/Hero'
-import AboutResume from '../components/AboutResume'
-import LikeToDo from '../components/LikeToDo'
-import ListWorksSkills from '../components/ListWorksSkills'
+import Hero from '../components/Hero/Hero'
+import AboutResume from '../components/AboutResume/AboutResume'
+import LikeToDo from '../components/LikeToDo/LikeToDo'
+import ListWorksSkills from '../components/ListWorksSkills/ListWorksSkills'
 
 function HomePage(): JSX.Element {
   return (
